Guessing Game Challenge
Write a program that picks a random integer from 1 to 100, and has players guess the number. The rules are:

If a player's guess is less than 1 or greater than 100, say "OUT OF BOUNDS"

On all turns, if a guess is

closer to the number than the previous guess return "Closer!"
farther from the number than the previous guess, return "Farther!"
When the player's guess equals the number, tell them they've guessed correctly and how many guesses it took!