Omar's family
file1 named omar.py that Generate a list with 15 Random Number between Range 10 -> 50 and save it as a list with name omar_list. file 2 named abo-omar.py that calls omar.py and add 5 to each element in omar's list and call it abo_omar_list. file 3 named om-omar.py that changes everything her family did by asking the user to give a number between 100 -> 200 and use it to modify the Range of omar_list and call abo-omar.py and name the output list as om_omar_list .